Cosmetic fix to a cast.  Should use long rather than int.
git-svn-id: svn://svn.code.sf.net/p/polyml/code/trunk/polyml@858 ae7d391e-3f74-4a2b-a0db-a8776840fd2a

+++ libpolyml/x86_dep.cpp
@@ -2023,7 +2023,7 @@ unsigned char *X86Dependent::BuildCallback(TaskData *taskData, int cbEntryNo, Ha
     cbEntryNo = cbEntryNo >> 8;
     *p++ = cbEntryNo & 0xff;
     /* The call is PC relative so we have to subtract the address of the END of the call instruction. */
-    cbAddr -= (int)p + 5; /* The instruction is 5 bytes long. */
+    cbAddr -= (long)p + 5; /* The instruction is 5 bytes long. */
     *p++ = 0xE8;    /* call cbAddr */
     *p++ = cbAddr & 0xff;
     cbAddr = cbAddr >> 8;
